feat(cspp): Now have CSPP_BASE_URL plumbed
Previously the application had no knowledge of where it was running or how so
it couldn't send canonical URIs to a user. Now we read from an ENV var to do
that. The var is CSPP_BASE_URL. IF that env var is not set, it falls back to
just using relaitve links.

Fixes #41