fix(cspp): Fix PORT and BASE_URL behavior
Originally, if `CSPP_PORT` and the port found in `CSPP_BASE_URL` weren't the
same, it would service on `CSPP_PORT` but link out and log with
`CSPP_BASE_URL`. THis was bad. To fix this, we now honor the port found in
`CSPP_BASE_URL` at all times, unless it's not set (e.g. it's just a string like
`http://example.com/cspp`) and `CSPP_PORT` is set. Then it will use `CSPP_PORT`
for all things. If neither is specified, it's just port 8080.