feat(hubot-wisdom): Add a wisdom hubot module
Adds a new hubot modules to enable quotes. This follows the format

"a quote I would like to remember" -- author

When that happens, hubot will respond with a :quote: emoji and that is the
indicator that the quote is now in the brain. To recall a random quote (ala
fortune), you simply run `hubot wisdom` and it will provide a random one. The
more quotes, the better!

closes #77