feat(zn): Initial zoom application layer API client
This is the first pass with limited functionality of having the application
also be optionally a client to zoom web api and not just a webhook receiver and
processor. Calling the API enables s things like getting the meeting join link
with a passcode or enumerating who is in the meeting.